+++ Changelog.md
@@ -2,6 +2,7 @@
 * Build orders are open with new `build_order.html` page.
 * Clean the game JS files to remove the BO content.
 * Add "Submit your build order" button.
+* Filtering with all added for opponent faction.
 
 # 2026.05.01
 * Script 'update_build_orders' supports `-r` option to remove BO.